[코드리뷰 반영] 애매하게 작성한 클래스명 변경
